fix union <:< support, add/move tests
