+++ izumi-reflect/izumi-reflect/src/test/scala_3/izumi/reflect/test/DottyProgressionTests.scala
@@ -11,16 +11,6 @@ class DottyProgressionTests extends TagAssertions {
 
   "dotty version" should {
 
-    // "does not fail on unresolved type parameters" in {
-    //   def badTag[T]: Tag[T] = Tag[T]
-    // }
-
-    // "does not fail on intersection/union of unresolved type parameters" in {
-    //   def badTag0[R, R0] = Tag[R with R0]
-    //   def badTag1[T, U]: Tag[T & U] = Tag[T & U]
-    //   def badTag2[T, U]: Tag[T | U] = Tag[T | U]
-    // }
-
     "fails to check subtyping when higher-kinds are involved" in {
       intercept[TestFailedException] {
         assertChild(LTT[FT2[IT2]], LTT[FT1[IT1]])
